Reset File Modified Time on Boot
When we boot a Dyno the slug is decompressed and we end up with a file last
modified time of when the Dyno was booted. This is not ideal as static files
that have not been modified end up getting a new timestamp on boot which
invalidates caches. Worse, if we have multiple Dynos scalled up to handle more
traffic we get conflicting last modified times and etags for static asssets.

This change forces all files in the webroot to have the same modified time,
that of when the slug was created.